+++ app/src/main/java/com/example/sstep/Checklist.java
@@ -9,6 +9,6 @@ public class Checklist extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_checklist);
+        setContentView(R.layout.checklist);
     }
 }
