+++ app/src/main/java/com/example/sstep/Checklist_detail.java
@@ -9,6 +9,6 @@ public class Checklist_detail extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_checklist_detail);
+        setContentView(R.layout.checklist_detail);
     }
 }
