+++ app/src/main/java/com/example/sstep/Find_id.java
@@ -8,6 +8,6 @@ public class Find_id extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_find_id);
+        setContentView(R.layout.find_id);
     }
 }
