+++ app/src/main/java/com/example/sstep/Find_password.java
@@ -9,6 +9,6 @@ public class Find_password extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_find_password);
+        setContentView(R.layout.find_password);
     }
 }
