+++ app/src/main/java/com/example/sstep/RegisterStore.java
@@ -9,6 +9,6 @@ public class RegisterStore extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_register_store);
+        setContentView(R.layout.register_store);
     }
 }
