+++ app/src/main/java/com/example/sstep/Search_Store.java
@@ -9,6 +9,6 @@ public class Search_Store extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_search_store);
+        setContentView(R.layout.search_store);
     }
 }
