+++ app/src/main/java/com/example/sstep/addSchedule.java
@@ -9,6 +9,6 @@ public class addSchedule extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_add_schedule);
+        setContentView(R.layout.add_schedule);
     }
 }
