+++ app/src/main/java/com/example/sstep/modifySchedule.java
@@ -9,6 +9,6 @@ public class modifySchedule extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_modify_schedule);
+        setContentView(R.layout.modify_schedule);
     }
 }
