+++ app/src/main/java/com/example/sstep/modifyStore.java
@@ -9,6 +9,6 @@ public class modifyStore extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_modify_store);
+        setContentView(R.layout.modify_store);
     }
 }
