+++ app/src/main/java/com/example/sstep/staff_workType.java
@@ -9,6 +9,6 @@ public class staff_workType extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_staff_work_type);
+        setContentView(R.layout.staff_work_type);
     }
 }
