+++ app/src/main/res/drawable/ychbox_black.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <selector xmlns:android="http://schemas.android.com/apk/res/android">
   <item android:drawable="@drawable/ychbox_checkedblack" android:state_checked="true"></item>
-  <item android:drawable="@drawable/ychbox_unchecked" android:state_checked="false"></item>
+  <item android:drawable="@drawable/ychbox_blackunchecked" android:state_checked="false"></item>
 </selector>
