+++ app/src/main/res/drawable/ychbox_blackunchecked.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="15dp"
+    android:height="15dp"
+    android:viewportWidth="14"
+    android:viewportHeight="14">
+  <path
+      android:pathData="M7,0C3.136,0 0,3.136 0,7C0,10.864 3.136,14 7,14C10.864,14 14,10.864 14,7C14,3.136 10.864,0 7,0ZM7,12.6C3.906,12.6 1.4,10.094 1.4,7C1.4,3.906 3.906,1.4 7,1.4C10.094,1.4 12.6,3.906 12.6,7C12.6,10.094 10.094,12.6 7,12.6Z"
+      android:fillColor="#DADADA"/>
+</vector>
