+++ app/src/main/res/drawable/ychbox_checkedblack.xml
@@ -1,6 +1,6 @@
 <vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="20dp"
-    android:height="20dp"
+    android:width="15dp"
+    android:height="15dp"
     android:viewportWidth="14"
     android:viewportHeight="14">
   <path
