+++ app/src/main/res/drawable/yicon_comment.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="21dp"
+    android:height="20dp"
+    android:viewportWidth="21"
+    android:viewportHeight="20">
+  <path
+      android:pathData="M10.028,0C8.715,0 7.415,0.259 6.201,0.761C4.988,1.264 3.886,2 2.957,2.929C1.082,4.804 0.028,7.348 0.028,10C0.019,12.309 0.819,14.549 2.288,16.33L0.288,18.33C0.149,18.471 0.055,18.649 0.018,18.843C-0.019,19.037 0.002,19.238 0.078,19.42C0.161,19.6 0.296,19.751 0.465,19.854C0.634,19.958 0.83,20.008 1.028,20H10.028C12.68,20 15.224,18.946 17.099,17.071C18.975,15.196 20.028,12.652 20.028,10C20.028,7.348 18.975,4.804 17.099,2.929C15.224,1.054 12.68,0 10.028,0ZM10.028,18H3.438L4.368,17.07C4.554,16.883 4.659,16.629 4.659,16.365C4.659,16.101 4.554,15.847 4.368,15.66C3.059,14.352 2.243,12.63 2.061,10.789C1.878,8.947 2.34,7.099 3.367,5.56C4.395,4.02 5.924,2.884 7.695,2.346C9.465,1.808 11.368,1.9 13.078,2.607C14.789,3.314 16.201,4.592 17.075,6.224C17.948,7.856 18.229,9.74 17.869,11.555C17.51,13.37 16.531,15.005 15.102,16.18C13.672,17.355 11.879,17.999 10.028,18Z"
+      android:fillColor="#737373"/>
+</vector>
