+++ app/src/main/res/drawable/yicon_image.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="145dp"
+    android:height="117dp"
+    android:viewportWidth="155"
+    android:viewportHeight="127">
+  <path
+      android:pathData="M144.36,0H10.64C7.82,0 5.11,1.09 3.12,3.03C1.12,4.97 0,7.6 0,10.34V116.66C0,119.4 1.12,122.03 3.12,123.97C5.11,125.91 7.82,127 10.64,127H144.36C147.18,127 149.89,125.91 151.88,123.97C153.88,122.03 155,119.4 155,116.66V10.34C155,7.6 153.88,4.97 151.88,3.03C149.89,1.09 147.18,0 144.36,0ZM10.64,8.86H144.36C144.77,8.86 145.15,9.02 145.44,9.29C145.72,9.57 145.88,9.95 145.88,10.34V89.76L123.49,68C122.5,67.04 121.33,66.28 120.04,65.76C118.75,65.24 117.37,64.98 115.97,64.98C114.57,64.98 113.19,65.24 111.9,65.76C110.61,66.28 109.43,67.04 108.45,68L92.17,83.82L57.67,50.28C56.68,49.32 55.51,48.56 54.22,48.04C52.93,47.52 51.54,47.26 50.15,47.26C48.75,47.26 47.37,47.52 46.08,48.04C44.79,48.56 43.61,49.32 42.63,50.28L9.12,82.85V10.34C9.12,9.95 9.28,9.57 9.56,9.29C9.85,9.02 10.23,8.86 10.64,8.86ZM9.12,116.66V95.38L49.07,56.55C49.21,56.41 49.38,56.3 49.56,56.23C49.75,56.15 49.95,56.12 50.15,56.12C50.35,56.12 50.55,56.15 50.73,56.23C50.92,56.3 51.08,56.41 51.23,56.55L114.6,118.14H10.64C10.23,118.14 9.85,117.98 9.56,117.71C9.28,117.43 9.12,117.05 9.12,116.66ZM144.36,118.14H127.5L98.62,90.08L114.89,74.27C115.03,74.13 115.2,74.02 115.38,73.94C115.57,73.87 115.76,73.83 115.96,73.83C116.17,73.83 116.36,73.87 116.55,73.94C116.73,74.02 116.9,74.13 117.04,74.27L145.91,102.32V116.66C145.91,116.86 145.87,117.05 145.79,117.24C145.72,117.42 145.6,117.58 145.46,117.72C145.31,117.86 145.14,117.96 144.95,118.04C144.77,118.11 144.57,118.14 144.36,118.14ZM91.18,42.83C91.18,41.37 91.62,39.94 92.46,38.72C93.29,37.51 94.48,36.56 95.87,36C97.26,35.45 98.78,35.3 100.26,35.58C101.73,35.87 103.08,36.57 104.15,37.6C105.21,38.64 105.93,39.95 106.23,41.39C106.52,42.82 106.37,44.3 105.79,45.65C105.22,47 104.25,48.15 103,48.96C101.75,49.78 100.28,50.21 98.77,50.21C96.76,50.21 94.83,49.43 93.4,48.05C91.98,46.66 91.18,44.78 91.18,42.83Z"
+      android:fillColor="#737373"/>
+</vector>
