+++ app/src/main/res/layout/activity_start.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".Start">
+    tools:context=".start.Start">
 
     <androidx.viewpager.widget.ViewPager
         android:id="@+id/start_pagers"
