+++ app/src/main/res/layout/find_id.xml
@@ -44,7 +44,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_marginStart="18dp"
-            android:layout_marginTop="15dp"
+            android:layout_marginTop="20dp"
             android:layout_marginEnd="18dp"
             android:background="@drawable/j_layout_background"
             android:orientation="vertical">
@@ -104,7 +104,7 @@
                         android:layout_marginBottom="20dp"
                         android:backgroundTint="#B9B9B9"
                         android:hint="01012345678"
-                        android:inputType="textPersonName"
+                        android:inputType="phone|number"
                         android:textColor="#000000"
                         android:textColorHint="#DCDCDC"
                         android:textSize="13sp" />
