+++ app/src/main/res/layout/find_password.xml
@@ -45,7 +45,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_marginStart="18dp"
-            android:layout_marginTop="15dp"
+            android:layout_marginTop="20dp"
             android:layout_marginEnd="18dp"
             android:background="@drawable/j_layout_background"
             android:orientation="vertical">
@@ -104,7 +104,7 @@
                         android:layout_marginTop="5dp"
                         android:backgroundTint="#B9B9B9"
                         android:hint="01012345678"
-                        android:inputType="textPersonName"
+                        android:inputType="phone|number"
                         android:textColor="#000000"
                         android:textColorHint="#DCDCDC"
                         android:textSize="13sp" />
@@ -151,7 +151,7 @@
                     android:layout_marginTop="5dp"
                     android:backgroundTint="#B9B9B9"
                     android:hint="영문, 숫자, 특수문자 조합 6자 이상"
-                    android:inputType="textPersonName"
+                    android:inputType="textPassword"
                     android:textColor="#000000"
                     android:textColorHint="#DCDCDC"
                     android:textSize="13sp" />
@@ -180,7 +180,7 @@
                     android:layout_marginTop="5dp"
                     android:backgroundTint="#B9B9B9"
                     android:hint="새 비밀번호 확인"
-                    android:inputType="textPersonName"
+                    android:inputType="textPassword"
                     android:textColor="#000000"
                     android:textColorHint="#DCDCDC"
                     android:textSize="13sp" />
