+++ app/src/main/res/layout/fragment_start1.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Start1">
+    tools:context=".start.Start1">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
