+++ app/src/main/res/layout/fragment_start2.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Start2">
+    tools:context=".start.Start2">
 
     <!-- TODO: Update blank fragment layout -->
