+++ app/src/main/res/layout/login.xml
@@ -30,10 +30,10 @@
             android:id="@+id/logIn_idEt"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:backgroundTint="#000000"
+            android:backgroundTint="#737373"
             android:ems="10"
             android:hint=" 아이디"
-            android:inputType="textPersonName"
+            android:inputType="text"
             android:textColor="#000000"
             android:textColorHint="#616161"
             android:textSize="15sp" />
@@ -43,10 +43,10 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_marginTop="16dp"
-            android:backgroundTint="#000000"
+            android:backgroundTint="#737373"
             android:ems="10"
             android:hint=" 비밀번호 (영문,숫자 조합 6자 이상)"
-            android:inputType="textPersonName"
+            android:inputType="textPassword"
             android:textColor="#000000"
             android:textColorHint="#616161"
             android:textSize="15sp" />
