+++ app/src/main/res/layout/register_store.xml
